Fixed #26401 -- Allowed use of auth logic without auth in INSTALLED_APPS.
