+++ django/contrib/auth/context_processors.py
@@ -1,3 +1,6 @@
+from django.contrib.auth.base_user import AnonymousUser
+
+
 # PermWrapper and PermLookupDict proxy the permissions system into objects that
 # the template system can understand.
 
@@ -57,7 +60,6 @@ def auth(request):
     if hasattr(request, 'user'):
         user = request.user
     else:
-        from django.contrib.auth.models import AnonymousUser
         user = AnonymousUser()
 
     return {
